+++ NEWS.md
@@ -1,5 +1,9 @@
 # Rsequoia2 (development version)
 
+## v0.0.10
+- further check when reading matrice to avoid empty rows in insee, section or 
+number field
+
 ## v0.0.9
 - bug fix for mnhn layer : more robust for loop and update available layers